pyproject.toml
[tool.poetry]
authors = ["Oleksandr Korol <oleksandr.korol@coaxsoft.com>"]
classifiers = [
  "Framework :: Django :: 3.2",
  "Framework :: Django :: 4.0",
  "Framework :: Django :: 4.1",
  "Framework :: Django :: 4.2",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: BSD License",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Framework :: Django",
]
description = "Package to handle various payments provider inside your drf project"
license = "MIT"
name = "drf-payments"
packages = [{include = "drf_payments"}]
readme = "README.md"
version = "0.0.2"
[project.urls]
Documentation = "https://coaxsoft.github.io/drf-payments/"
Repository = "https://github.com/coaxsoft/drf-payments"

[tool.poetry.dependencies]
Django = ">3.2.0"
braintree = "^4.20.0"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^7.1.0"}
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.5"
python = ">=3.8"
stripe = "^5.4"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.270"

[tool.poetry.group.test.dependencies]
coverage = {extras = ["toml"], version = "^7.2.7"}
tox = "^4.6.3"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.3"
mkdocs-awesome-pages-plugin = "^2.9.1"
mkdocs-material = "^9.1.17"
mkdocstrings = {extras = ["python"], version = "^0.22.0"}

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
# Ruff settings
[tool.ruff]
ignore = ["RUF005"]
line-length = 120
select = ["E", "F", "W", "I", "C4", "Q", "RUF", "PIE", "COM"]
# Exclude a variety of commonly ignored directories.
exclude = [
  ".git",
  ".hg",
  ".mypy_cache",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "migrations",
  "dist",
  "venv",
  ".venv",
  ".conf",
  ".data",
  ".env",
  "asgi.py",
]
per-file-ignores = {}
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
target-version = "py38"
[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10

# Black settings

[tool.black]
line-length = 120
target-version = ['py310']

[tool.coverage.run]
branch = true
omit = [
  ".venv/*",
  "*/migrations/*",
  "*/admin.py",
  "*/apps.py",
  "manage.py",
  "*/settings.py",
  "*/asgi.py",
  "*/wsgi.py",
  "*/tests.py",
  "example/*",
  "shop/*",
]

[tool.coverage.report]
show_missing = true
skip_empty = true
[tool.tox]
legacy_tox_ini = """
    [tox]
    env_list =
        py{38,39,310}-django{320}
        py{38,39,310,311}-django{400}
        py{38,39,310,311}-django{410}
        py{38,39,310,311}-django{420}


    [testenv]
    allowlist_externals = 
        make
    commands = 
        make tests
    deps = 
        coverage
        tomli
        braintree
        stripe
        djangorestframework
        django-phonenumber-field[phonenumberslite]
        django320: Django>=3.2,<3.3
        django400: Django>=4.0,<4.1
        django410: Django>=4.1,<4.2
        django420: Django>=4.2,<5.0

"""
